requirements.txt
opencv-python~=4.7.0.68
numpy~=1.24.1
matplotlib~=3.6.3
pytesseract~=0.3.10
Pillow~=9.4.0
Flask~=2.2.2